Changed platform collision to use platform width and height for
-Distance calculation for collision detection now uses platform width
and height instead of player characteristics. This may result in
additional active platforms however it will resolve issues with
platforms that are taller than they are wide.

-Now using JQuery locally so I don't have to be online to program.

-Started working on left-side collision detection… it needs some work.

-Started to re-work collision detection from the bottom to avoid being
able to jump through corners or be pushed through the floor below.

-Quick and dirty fix to prevent player from falling a few pixels
through the top of platforms. Will probably change to do some
calculations during free-fall later on.